increase groundDistance
Increasing the size of groundDistance seems to keep the player grounded whenever they are climbing slopes.

note:  I don't know why, but even if the radius of the ground check sphere is outside the player collider, it won't register you as grounded unless you put the transform very close to the floor. No idea why.