Fix Slope Climbing
So turns out the groundDistance is an absolute value. I had set the scale of  PlayerParent to 3, I needed to be setting my groundDistance to be 3 times larger than what I wanted in terms of the TPPlayer (the actual player element).

The camera is a bit shaky now, not sure if that's just because of the bumpy road and I've just never noticed it, or if it's actually caused by what I did.